+++ packages/react-widgets/src/models/HistogramModel.js
@@ -1,6 +1,6 @@
-import { executeSQL } from '@carto/react-api';
+import { _executeModel } from '@carto/react-api';
 import { Methods, executeTask } from '@carto/react-workers';
-import { formatTableNameWithFilters, normalizeObjectKeys, wrapModelCall } from './utils';
+import { normalizeObjectKeys, wrapModelCall } from './utils';
 
 export function getHistogram(props) {
   return wrapModelCall(props, fromLocal, fromRemote);
@@ -21,36 +21,19 @@ function fromLocal(props) {
 
 // From remote
 async function fromRemote(props) {
-  const { source, ticks, abortController } = props;
-  const { credentials, connection } = source;
+  const { source, abortController, ...params } = props;
 
-  const query = buildSqlQueryToGetHistogram(props);
+  const { column, operation, ticks } = params;
 
-  const data = await executeSQL({
-    credentials,
-    query,
-    connection,
+  const data = await _executeModel({
+    model: 'histogram',
+    source,
+    params: { column, operation, ticks },
     opts: { abortController }
-  }).then(normalizeObjectKeys);
+  }).then((res) => normalizeObjectKeys(res.rows));
 
   const result = Array(ticks.length + 1).fill(0);
   data.forEach(({ tick, value }) => (result[tick] = value));
 
   return result;
-}
-
-function buildSqlQueryToGetHistogram(props) {
-  const { column, operation, ticks } = props;
-
-  const caseTicks = ticks.map((t, index) => `WHEN ${column} < ${t} THEN ${index}`);
-  caseTicks.push(`ELSE ${ticks.length}`);
-
-  const selectValueClause = `${operation}(${column}) as value`;
-  const selectTickClause = `CASE ${caseTicks.join(' ')} END as tick`;
-
-  const tableName = formatTableNameWithFilters(props);
-
-  const subQuery = `SELECT ${selectTickClause}, ${column} FROM ${tableName}`;
-
-  return `SELECT tick, ${selectValueClause} FROM (${subQuery}) q GROUP BY tick`;
 }